fix: move.Meta is no longer a pointer
ill be honest, don't know why this was a pointer. my guess is that I
wanted it to be an optional type. if thats the case then this is
offically not my fault and actually its a problem with Go. though in all
seriousness i miss optional types. Go would be like at least 10 times
better with an explicit optional/nullable type.

-- Rant Over --